refactor: init 파일 활용을 통한 import 경로 수정 및 의존성 추가
