+++ qr/serializers/__init__.py
@@ -1,0 +1,2 @@
+from qr.serializers.place_serializer import PlaceSerializer
+from qr.serializers.qr_image_serializer import QrImageSerializer
