+++ qr/views/__init__.py
@@ -1,0 +1,2 @@
+from qr.views.place_view import PlaceView
+from qr.views.qr_image_view import QrImageView
