+++ qr/views/place_view.py
@@ -1,7 +1,7 @@
 from rest_framework.response import Response
 from rest_framework import status, exceptions
 from rest_framework.views import APIView
-from qr.services.place_service import create_place
+from qr.services import create_place
 
 class PlaceView(APIView):
     """
