+++ requirements.txt
@@ -1,3 +1,4 @@
+annotated-types==0.5.0
 asgiref==3.7.2
 async-timeout==4.0.2
 attrs==23.1.0
@@ -30,21 +31,27 @@ jmespath==1.0.1
 msgpack==1.0.5
 oauthlib==3.2.2
 packaging==23.1
+Pillow==10.0.0
 pyasn1==0.5.0
 pyasn1-modules==0.3.0
 pycparser==2.21
+pydantic==2.1.1
+pydantic_core==2.4.0
 PyJWT==2.8.0
 PyMySQL==1.1.0
 pyOpenSSL==23.2.0
+pypng==0.20220715.0
 python-dateutil==2.8.2
 python-decouple==3.8
 python3-openid==3.2.0
 pytz==2023.3
 PyYAML==6.0.1
+qrcode==7.4.2
 redis==4.6.0
 requests==2.31.0
 requests-oauthlib==1.3.1
 s3transfer==0.6.1
+segno==1.5.2
 service-identity==23.1.0
 six==1.16.0
 sqlparse==0.4.4
